refactoring strong script
